Completed Feature! see description
Cleaned up bugs with the algorithm, made edits to BrowseStationsActivity -> Main  Activity -> Station class to correctly log tags and likes data over to Parse on radio-browser selection. Then that 'tags' data is used as a query parameter to parse, while 'likes' data is used to calculate a weight in the algorithm score for a given nearby station.